+++ posthog/src/main/java/com/posthog/internal/PostHogSendCachedEventsIntegration.kt
@@ -129,7 +129,7 @@ internal class PostHogSendCachedEventsIntegration(
         storagePrefix: String?,
         endpoint: PostHogApiEndpoint,
     ) {
-        storagePrefix.let {
+        storagePrefix?.let {
             val dir = File(it, config.apiKey)
 
             if (!dir.existsSafely(config)) {