+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## Next
 
 - chore: Session Replay - GA
-- fix: sending cached events null check ([#214](https://github.com/PostHog/posthog-flutter/pull/214))
+- fix: sending cached events null check ([#218](https://github.com/PostHog/posthog-flutter/pull/218))
 
 ## 3.10.0 - 2025-01-07