+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Next
 
+## 3.11.0 - 2025-01-27
+
 - chore: Session Replay - GA
 - fix: session replay and auto capture works with 'with' method ([#217](https://github.com/PostHog/posthog-flutter/pull/217))
 - fix: sending cached events null check ([#218](https://github.com/PostHog/posthog-flutter/pull/218))